+++ CHANGELOG.md
@@ -10,6 +10,9 @@ Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to
 ## [Unreleased]
 
 ### Added
+
+- Added new `functions` module for operations on NumPy arrays.
+
 ### Changed
 ### Deprecated
 ### Removed